Add and use Atkinson Hyperlegible and Fira Code fonts
These are being added per the discussion at:
https://lists.riscv.org/g/sig-documentation/message/343

The Atkinson Hyperlegible font is being used as the
default font for the document. Monospace and source
code listings are using Fira Code. Both fonts are
freely available under a SIL OpenFonts license and
those license documents are included. The TTF files
are currently used, but SVG, WOFF, and WOFF2 are also
added for future use with web based display platforms.